Enemy Pool 생성 및 좀비 죽음 처리

+++ ZomBoogie_Unity/Assets/MainZomBoogie/Boogie/BaseStatus.cs
@@ -7,8 +7,8 @@ using System.Numerics;
 [System.Serializable]
 public class BaseStatus
 {
-    public float mHealth;
-    public float mAttack;
+    public int mHealth;
+    public int mAttack;
     public float mSpeed;
 }
 [System.Serializable]
